==> beijing_traffic/01_road_feature_matrix_full.ipynb <==
# Tensorflow2.0 based code

import tensorflow as tf

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
%pylab inline

path = '/home/blu/workspace/graduate_project/data/'

## 一共有45148条路，每条路5个特征

road_network = pd.read_table(path+'road_network_sub-dataset.v2')
road_network.head()

road_network.shape

## Feature engineering

- 连续变量

item | min | max | mean | std | unit
:-:|:-:|:-:|:-:|:-:|:-:
length | 0.002 | 0.492 | 0.0588 | 0.063 | 米


- 离散变量

item | value | meanings |
:-:|:-:|:-:
width | {15,30,55} | 15：小于等于3m 30：3m至5m之间 55：5.5m至13m之间
direction | {1,2,3} | 1：双向 2：单向（s->e）3：单向（e->s）
speedclass | {2,3,4,5,6,7,8} | 实际上是没有1的，速度class越小越快
lanenum | {1,2,3} | 1：单车道 2：双或三车道 3：多车道

def width_label_change(x):
    if(x['width'] == 15):
        return 0
    elif(x['width'] == 30):
        return 1
    else:
        return 2
rr = road_network.apply(lambda x: width_label_change(x), axis=1)
road_network["width_label"] = rr

def direction_label_change(x):
    x['direction'] = x['direction'] - 1
    return x['direction']
rr = road_network.apply(lambda x: direction_label_change(x), axis=1)
road_network["direction"] = rr

def speedclass_label_change(x):
    x['speedclass'] = x['speedclass'] - 2
    return x['speedclass']
rr = road_network.apply(lambda x: speedclass_label_change(x), axis=1)
road_network["speedclass"] = rr

def lanenum_label_change(x):
    x['lanenum'] = x['lanenum'] - 1
    return x['lanenum']
ll = road_network.apply(lambda x: lanenum_label_change(x), axis=1)
road_network["lanenum"] = ll

road_network.head()

width_oh = tf.keras.utils.to_categorical(road_network['width_label'], num_classes = 3)
print(width_oh[0:5])

direction_oh = tf.keras.utils.to_categorical(road_network['direction'], num_classes = 3)
print(direction_oh[0:5])

speedclass_oh = tf.keras.utils.to_categorical(road_network['speedclass'], num_classes = 7)
lanenum_oh =  tf.keras.utils.to_categorical(road_network['lanenum'], num_classes = 3)

road_network.describe()['length']

def length_label_change(x):
    length = int(x['length']*100)
    if(length <= 1):
        return 0
    elif(length >= 2 and length <= 3):
        return 1
    elif(length >= 4 and length <= 7):
        return 2
    else:
        return 3
rr = road_network.apply(lambda x: length_label_change(x), axis=1)
road_network["length_label"] = rr

length_oh =  tf.keras.utils.to_categorical(road_network['length_label'], num_classes = 4)

print("width:", width_oh.shape)
print("direction:", direction_oh.shape)
print("length:", length_oh.shape)
print("speedclass:", speedclass_oh.shape)
print("lanenum: ", lanenum_oh.shape)


feature_oh = tf.concat((width_oh, direction_oh, length_oh, speedclass_oh, lanenum_oh),axis=1)
feature_oh.shape

feature_oh.numpy()

feature_matrix = pd.DataFrame(feature_oh.numpy())
feature_matrix.head()

feature_matrix["link_id"] = road_network["link_id"]
feature_matrix.head()

cols = list(feature_matrix)
cols.insert(0,cols.pop(cols.index('link_id')))
feature_matrix = feature_matrix.loc[:,cols]
feature_matrix.head()

save_path = '/home/blu/workspace/graduate_project/STGCN/beijing_traffic/01_road_feature_matrix_full.npy'
np.save(save_path, feature_matrix.values)
