==> beijing_traffic/02_select_node_from_LargeMap.ipynb <==
!pip3 install tqdm

import numpy as np
import pandas as pd
from tqdm import tqdm

df = pd.read_table('link_gps.v2', names = ['link_id', 'longtitude', 'latitude'])

df.shape[0]

#boundary_1右上，boundary_2左下
boundary_1 = [116.46812,39.929101]
boundary_2 = [116.457053,39.919362]

left_x = boundary_2[0]
right_x = boundary_1[0]
up_y = boundary_1[1]
down_y = boundary_2[1]

x = df['longtitude'][5]
y = df['latitude'][5]
if(x >= left_x and x <= right_x):
    if(y >= down_y and y <= up_y):
        print('True')
    else:
        print('False')
else:
    print('False')

selected_id = []
for i in tqdm(range(df.shape[0])):
    x = df['longtitude'][i]
    y = df['latitude'][i]
    if(x >= left_x and x <= right_x):
        if(y >= down_y and y <= up_y):
            selected_id.append(df['link_id'][i])

len(selected_id)

np_selected = np.array(selected_id)

np.save("selected_id.npy", np_selected)
