==> beijing_traffic/03_select_road_map_feature.ipynb <==
import numpy as np
import pandas as pd

selected_id = np.load("selected_id.npy")

road_feature = np.load('01_road_feature_matrix_full.npy')

road_feature = pd.DataFrame(road_feature)

road_feature[0] = road_feature[0].astype(int)

road_feature.head()

select_road_feature = road_feature[road_feature[0].isin(selected_id)]

print('len(selected_id) = ', len(selected_id))
print('selected_road_feature shape = ', select_road_feature.shape)

select_road_feature_arr = select_road_feature.values

np.save('select_road_feature_arr.npy', select_road_feature_arr)
