==> beijing_traffic/04_select_speed_data.ipynb <==
import pandas as pd
import numpy as np

path = 'D:/2019-2020-1/BaiduTraffic/traffic_speed_sub-dataset.v2/traffic_speed_sub-dataset/traffic_speed_sub-dataset_02.v2'
speed_df = pd.read_table(path, names = ['link_id', 'time', 'speed'], sep = ',')
speed_df.head()

for i in range(2,10):
    path = 'D:/2019-2020-1/BaiduTraffic/traffic_speed_sub-dataset.v2/traffic_speed_sub-dataset/traffic_speed_sub-dataset_0%d.v2'%(i)
    print(path)
    speed_df = pd.read_table(path, names = ['link_id', 'time', 'speed'], sep = ',')
    selected_speed = speed_df[speed_df['link_id'].isin(selected_id)]
    c = np.concatenate((c,selected_speed), axis=0)
    print(c.shape)

speed_df.shape

selected_id = np.load("02_selected_id.npy")

selected_speed_2 = speed_df[speed_df['link_id'].isin(selected_id)]

a = selected_speed.values

a

b = selected_speed_2.values

c = np.concatenate((a,b), axis=0)

c.shape

np.save("selected_speed_data.npy", c)

df = pd.DataFrame(c)

df.columns = ['link_id', 'time', 'speed']

df['link_id'] = df['link_id'].astype(int)

df['time'] = df['time'].astype(int)

df.head()

c = df.values

c
