==> beijing_traffic/05_generate_training_data.ipynb <==
from __future__ import absolute_import, division, print_function, unicode_literals

import collections
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd

import tensorflow as tf

from tensorflow.keras import layers

from tqdm import tqdm

## 处理数据

一共206条路，每条路有5856个time slot数据，时间间隔是15分钟，我们用过去的12组数据，预测未来的一步

speed_np = np.load('/home/blu/workspace/graduate_project/STGCN/beijing_traffic/04_selected_speed_data.npy')

speed_df = pd.DataFrame(speed_np)

speed_df.columns = ['link_id', 'time', 'speed']
speed_df['time'] = speed_df['time'].astype(int)
speed_df.head()

speed_df = speed_df.drop(['link_id'], axis=1)   #删除link_id

len(speed_df['time'].unique())

df_np = []
for i in range(len(speed_df['time'].unique())):
    df_arr = speed_df[speed_df['time']==i]['speed'].values
    df_np.append(df_arr)
df_np = np.stack(df_np, axis=0)

df = pd.DataFrame(df_np)
df.head()

df.shape

x = []
y = []
x_offset = 12
y_offset = 1

index_max = (df.shape[0]) - x_offset - y_offset
for i in tqdm(range(index_max)):
    x_array = df[i : i + x_offset].values
    x_array = np.array(x_array).flatten()
    y_array = df[i + x_offset : i + x_offset + y_offset].values
    y_array = np.array(y_array).flatten()
    
    x.append(x_array.tolist())
    y.append(y_array.tolist())

x = np.stack(x, axis=0)    
y = np.stack(y, axis=0)

print("x.shape = ", x.shape, " y.shape = ", y.shape)

x与y为所有数据，按照训练集70%，验证集10%，测试集20%的比例进行划分

num_samples = x.shape[0]
num_test = round(num_samples * 0.2)
num_train = round(num_samples * 0.7)
num_val = num_samples - num_test - num_train

x_train, y_train = x[:num_train], y[:num_train]
x_val, y_val = (
        x[num_train: num_train + num_val],
        y[num_train: num_train + num_val],
    )
x_test, y_test = x[-num_test:], y[-num_test:]

np.save("/home/blu/workspace/graduate_project/STGCN/beijing_traffic/x_train.npy", x_train)
np.save("/home/blu/workspace/graduate_project/STGCN/beijing_traffic/y_train.npy", y_train)

np.save("/home/blu/workspace/graduate_project/STGCN/beijing_traffic/x_test.npy", x_test)
np.save("/home/blu/workspace/graduate_project/STGCN/beijing_traffic/y_test.npy", y_test)
np.save("/home/blu/workspace/graduate_project/STGCN/beijing_traffic/x_val.npy", x_val)
np.save("/home/blu/workspace/graduate_project/STGCN/beijing_traffic/y_val.npy", y_val)
